+++ doc/whatsnew.rst
@@ -6,8 +6,10 @@ What's new
    :backlinks: none
    :depth: 1
 
-.. Next release
-.. ============
+Next release
+============
+
+- Add :func:`.testing.assert_units` (:pull:`58`).
 
 v1.9.2 (2022-03-03)
 ===================